refactor: make server/grpc package
fixes import cycle
